+++ app/models/ConjuredItem.ts
