+++ test/golden-master-text-test.ts
@@ -1,7 +1,5 @@
 import { GildedRose } from '../app/gilded-rose';
-import { ItemFactory } from "@/ItemFactory";
-import { BaseItem } from "@/BaseItem";
-
+import { ItemFactory } from "../app/helpers/ItemFactory";
 const items = [
   ItemFactory.createItem("+5 Dexterity Vest", 10, 20), //
   ItemFactory.createItem("Aged Brie", 2, 0), //
